# openapi/main.yaml
openapi: 3.1.0
info:
  title: AI裁判官
  summary: AIが補助してゲームみたいに楽しく揉め事を解決するWebアプリのAPI
  version: 1.0.1
paths:
...
  /trial/game_state/:
    post:
      tags: [Trial]
      summary: ゲームの初期進行状態を取得する
      description: |
        プロジェクター，スマホ用 <br>
        ゲームの初期進行状態を取得するエンドポイント。<br>
        - 2つのQRコードを表示するフェーズ (show_two_qr_codes),
        - 1つのQRコードを表示するフェーズ (show_one_qr_codes), 
        - 原告の主張，被告の主張，裁判官の暫定的な判決を表示するフェーズ (show_first_claim_and_judge),
        - 話し合いのフェーズ (discussion)
        - 最後の主張と判決を表示するフェーズ (show_final_claim_and_judge)
        を指示する．
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrialGameState'
      responses:
        '200':
          description: ゲームの初期進行状態取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrialGameStateResponse'
...
  /trial/claim/:
    post:
      tags: [Trial]
      summary: 主張を作成する
      description: |
        スマホ用 <br>
        原告、被告がQRコードを読み取り、名前を入力した後、最初の主張で入力した内容をPOSTするエンドポイント.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimCreate'
      responses:
        '201':
          description: 主張作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimCreateResponse'
    patch:
      tags: [Trial]
      summary: 主張を更新する
      description: |
        スマホ用 <br>
        裁判内容に対する主張を更新するエンドポイント。<br>
        ゲーム開始後は、原告、被告が主張を更新することができ、その時はこのPATCHを叩いて更新する.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimUpdate'
      responses:
        '200':
          description: 主張更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimUpdateResponse'
...
components:
  schemas:
...
    claim:
      type: string
      maximum: 100
      examples:
        - "お腹が空いていたとしても人のプリンを食べるのは重罪"
...
    game_state:
      type: string
      enum: [show_two_qr_codes, show_one_qr_codes, show_first_claim_and_judge, discussion, show_final_claim_and_judge]
      examples:
        - discussion
    id:
      type: string
      format: uuid
      examples:
        - 123e4567-e89b-12d3-a456-426614174000
...
    ClaimCreate:
      type: object
      properties:
        trial_id:
          $ref: '#/components/schemas/id'
        player_id:
          $ref: '#/components/schemas/id'
        claim:
          $ref: '#/components/schemas/claim'
      required:
        - trial_id
        - player_id
        - claim
    ClaimCreateResponse:
      type: object
      properties:
        trial_id:
          $ref: '#/components/schemas/id'
        player_id:
          $ref: '#/components/schemas/id'
        claim:
          $ref: '#/components/schemas/claim'
...
    ClaimUpdate:
      type: object
      properties:
        trial_id:
          $ref: '#/components/schemas/id'
        player_id:
          $ref: '#/components/schemas/id'
        claim:
          $ref: '#/components/schemas/claim'
      required:
        - trial_id
        - player_id
        - claim
    ClaimUpdateResponse:
      type: object
      properties:
        trial_id:
          $ref: '#/components/schemas/id'
        player_id:
          $ref: '#/components/schemas/id'
        claim:
          $ref: '#/components/schemas/claim'
...
    TrialGameState:
      type: object
      properties:
        trial_id:
          $ref: '#/components/schemas/id'
    TrialGameStateResponse:
      type: object
      properties:
        trial_id:
          $ref: '#/components/schemas/id'
        game_state:
          $ref: '#/components/schemas/game_state'
...